Base/CoreSvc/Controllers/CommunicationDefinitionController.cs
using System.Threading.Tasks;
using CoreSvc.Filters;
using CoreType.Types;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using CoreType.DBModels;
using CoreSvc.Services;

namespace CoreSvc.Controllers
{
    [Authorize]
    [DefaultRoute]
    [Resources("CommunicationDefinition_Res")]
    public class CommunicationDefinitionController : BaseController
    {
        private readonly CommunicationDefinitionService _mainService = new CommunicationDefinitionService();

        [HttpPost]
        [ClaimRequirement(ActionType.List)]
        public async Task<ResponseWrapper<PaginationWrapper<CommunicationDefinition>>> List([FromBody] RequestWithPagination<CommunicationDefinition> request)
        {
            return await _mainService.ListAsync(request);
        }

        [HttpPost]
        [ClaimRequirement(ActionType.GetRecord)]
        public async Task<ResponseWrapper<CommunicationDefinition>> Get([FromBody] CommunicationDefinition request)
        {
            return await _mainService.GetAsync(request);
        }

        [HttpPost]
        [ClaimRequirement(ActionType.Insert)]
        public async Task<ResponseWrapper<CommunicationDefinition>> Insert([FromBody] CommunicationDefinition request)
        {
            return await _mainService.SaveAsync(request);
        }

        [HttpPost]
        [ClaimRequirement(ActionType.Update)]
        public async Task<ResponseWrapper<CommunicationDefinition>> Update([FromBody] CommunicationDefinition request)
        {
            return await _mainService.SaveAsync(request);
        }

        [HttpPost]
        [ClaimRequirement(ActionType.Import)]
        public Task<ResponseWrapper> BulkSave([FromBody] RequestWithExcelData<CommunicationDefinition> request)
        {
            return _mainService.BulkSaveAsync(request);
        }

        [HttpPost]
        [ClaimRequirement(ActionType.Delete)]
        public async Task<ResponseWrapper<bool>> Delete([FromBody] CommunicationDefinition request)
        {
            return await _mainService.DeleteAsync(request);
        }
    }
}